test(setupcheck): reorganize test mocks into dedicated namespace
Move ExecMock and FileSystemMock from Unit/SetupCheck to the Mock/ directory to centralize test doubles and follow standard testing practices.

Consolidate function overrides (file_exists, exec) from multiple test files into a single functions.php file to improve maintainability and reduce duplication.

This refactoring addresses the reviewer's request by:
- Placing mocks in the proper namespace (OCA\Libresign\Tests\Mock) before the function overrides, aligning with PHP namespace fallback mechanics.
- Centralizing mock logic in dedicated, reusable classes.
- Making test setup cleaner and more maintainable.

Related issue: #6590
Type: Test
