--- tests/php/Unit/SetupCheck/Mock/ExecMock.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/**
- * SPDX-FileCopyrightText: 2026 LibreCode coop and contributors
- * SPDX-License-Identifier: AGPL-3.0-or-later
- */
-
-namespace OCA\Libresign\Tests\Unit\SetupCheck\Mock;
-
-class ExecMock {
-	public static array $commands = [];
-}
