--- tests/php/Unit/SetupCheck/Mock/FileSystemMock.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace OCA\Libresign\Tests\Unit\SetupCheck\Mock;
-
-class FileSystemMock {
-	public static array $files = [];
-}
