+++ tests/php/bootstrap.php
@@ -17,10 +17,6 @@ if (!defined('PHPUNIT_RUN')) {
 require_once __DIR__ . '/../../../../lib/base.php';
 require_once __DIR__ . '/../../../../tests/autoload.php';
 
-require_once __DIR__ . '/Unit/SetupCheck/Mock/FileSystemMock.php';
-require_once __DIR__ . '/Unit/SetupCheck/Mock/ExecMock.php';
-require_once __DIR__ . '/Unit/SetupCheck/Mock/functions.php';
-
 \OC::$composerAutoloader->addPsr4('Test\\', OC::$SERVERROOT . '/tests/php/lib/', true);
 \OC::$composerAutoloader->addPsr4('Tests\\', OC::$SERVERROOT . '/tests/php/', true);
 \OC::$composerAutoloader->addPsr4('OCA\\Libresign\\Tests\\', __DIR__, true);
